Make FenFeed possible to use, although simplistic and slow and sometimes buggy
- store list of subscriptions in a configuration file (in Turtle;
  the graph contains one node with type fenfeed:Configuration, which is
  connected to the conf information)

- remember which items have been read, also in the configurations file
  ("CONFIG fenfeed:read ITEM" -- argh, these triples are never removed
  even when the item disappears)

- show unread items, and channels containing unread items, in bold

The FenFeed code currently is a real mess, because the bolding is NOT
easy to do with the current lobs; it's POSSIBLE, but you need to string
lots of lobs together to achieve the intended effect. Need to think
about how to make that easier.

darcs-hash:20041206022120-2ee97-ba0a018a99fdb63dde36b64ef1736a23b07ddf88