Move CanvasCursor out of org.fenfire.Cursor; add adapter model that gives the panX or panY or zoom, given a Model containing a (Canvas or non-Canvas)Cursor.
CanvasCursor is a special type of cursor, not a generic one
(i.e. it applies only to some types of structures and views)
and shouldn't be in Cursor or in the general org.fenfire package.
Moved to org.fenfire.view because CanvasSpatialView is there.

darcs-hash:20050221154240-2ee97-72c339221a927ba5316c3f5e172e03df0ebba72e