Add a way to evaluate expressions when the user presses the key, before the command is complete
When pressing keys using the potions interface, the user builds up
an expression representing a command. When the expression is complete,
it's evaluated and the command is executed. However, say the user wants to
"connect (this node here) to (that node there)" -- they press Alt-T
or whatever for "this node," then go the other node, and press Alt-T again.

Clearly, the "this node" should be evaluated when the user presses Alt-T,
not when the command is complete.

So this patch adds this new functionality called a 'pattern' (for need
of a better name) that evaluates when the user presses the key, returning
a new Function that is then evaluated when the command is complete.

darcs-hash:20050223204211-2ee97-08af3791ecbffcb96819db94da4ab9d716d2ccdf