When creating a new cell, move the cursor onto it.
We used to have this elaborate code in FentwineKeyController
and its ancestors which would take care that things get
interpolated right in this case. I tried to port this to the
potions system, but I couldn't get it to work in a sensible way,
because it would have required involved communication between
the "make a new node" function and the "create a connection" command...

So I put a heuristic in ConnectionVobMatcher which checks
whether the old and new scenes' foci have the same key,
and if not, finds the old scene's focus's key in the new scene,
and uses that as the new scene's focus for the purpose of interpolation.

It won't always work, but I guess it's good enough for now.

darcs-hash:20050223231202-2ee97-3417f7c090d6ac33eb9ccf1b15152bece3effaf1