Radically simplify cursor architecture.
After actually trying out the scheme proposed on ff-dev, it didn't look
that great any more. ;-) Tried it out, made the code using it more complex
rather than simpler. So in the new code, Cursor is simply an object
holding a focused node model, spatial cursor model, text cursor model,
and Map of Rotation objects, and a couple of convenience functions
that change the cursor position in the usual patterns.

The interpolation problem that moving to the new structure solved
is back open, but we worked out a simple scheme on IRC to solve this
that I'll implement next.

darcs-hash:20050225203359-2ee97-42af48fb2083723c8397a52dc5d1e382569182e6