Allow the user to move on / edit literals.
Unlike the previous Fentwine implementation, don't treat literals
too specially. The old implementation used a special kind of NeighbourWheel
only for literals; the new implementation treats literals essentially
like other nodes, and handles only text editing specially.

The challenge with literals is that:

- In the RDF model, the literal "55" is a node different from
  the literal "5", and there is exactly one literal "55" and one literal "5"
  (a canvas:x "55"; b canvas:y "55" means that a and b are connected
  to the same literal.
- But in Fentwine, we want to *pretend* that literals are nodes that
  the user can move on and edit; i.e., the user moves on the literal "55",
  hits backspace, and "changes" it to the literal "5". (In "reality,"
  this removes one triple and adds another one, with a different literal
  as the object, but the user shouldn't see this.)
- Also, in Fentwine, we want to pretend that the "55"s in
  (a canvas:x "55") and (b canvas:y "55") are *different* nodes,
  one connected only to 'a' and one connected only to 'b' -- and "changing" one
  does not change the other.

The current scheme is that the text model that SimpleContentView uses
for literals has magic that changes the cursor and rotation
when the text is edited. Otherwise, we use the same code as for other nodes,
which means that all the connections of the literal are shown -- so we don't
get the third point above, but I guess that's not so important
and the new system actually reflects the RDF structure better, so
maybe it's nice. (If it turns out to be annoying, it actually shouldn't be
that hard to tweak the behavior so that no additional connections
of the literal are shown.)

darcs-hash:20050226014943-2ee97-09a1a0f30ca306a592760e11b4a55635988de1bf