New smushing architecture -- the UI code needs to know about smushing
In RDF, two different nodes can represent the same "resource"
(for example, the same person). In real-world data on the Web,
this frequently happens. For example, FOAF files normally use
bnodes (blank nodes, nodes without a URI) to represent people.
Two different FOAF files talking about the same person will thus
use different nodes (bnodes in different graphs are different
by definition) to represent the same person.

RDF and OWL make sure that the computer can know that two nodes
represent the same resource. For example, given

    _:foo foaf:mbox <mailto:foo@example.com>.
    _:bar foaf:mbox <mailto:foo@example.com>.

we can conclude that _:foo and _:bar represent the same person,
because foaf:mbox is defined this way:

    foaf:mbox rdf:type owl:inverseFunctionalProperty.

*Smushing* refers to merging two nodes that represent
the same resource; for example,

    _:foo foaf:mbox <mailto:foo@example.com>.
    _:foo foaf:name "Master Foo".

    _:bar foaf:mbox <mailto:foo@example.com>.
    _:bar sl:linkedTo _:blargh.

could become

    _:bar foaf:mbox <mailto:foo@example.com>.
    _:bar sl:linkedTo _:blargh.
    _:bar foaf:name "Master Foo".

I.e., _:bar has "replaced" _:foo -- all the connections
that used to be _:foo connections are now _:bar connections.

There's an alternative to smushing -- we could continue
to use both nodes:

    _:foo foaf:mbox <mailto:foo@example.com>.
    _:foo sl:linkedTo _:blargh.
    _:foo foaf:name "Master Foo".

    _:bar foaf:mbox <mailto:foo@example.com>.
    _:bar sl:linkedTo _:blargh.
    _:bar foaf:name "Master Foo".

But besides being wasteful, the bigger problem is that for us,
this would mean that the user would see two "Master Foos"
whenever looking at a resource linked to the master -- for example,
when looking at _:blargh, we would see it sl:linkedTo two nodes,
both of which say "Master Foo." So we have to smush.

Now we have the problem that if our cursor position is _:foo,
and _:foo gets merged into _:bar as in the smushing example,
_:foo is left without connections -- and we're suddenly looking
at an empty node without connections. Not nice ;-)

My original solution for this was to make the graph, when
asked for connections to _:foo, return the connections of _:bar--
but when asked for the connections of _:blargh, return only
the connection to _:bar, not the one to _:foo.

However, this broke bi-directionality, which *did* cause problems
in practice -- and left me wondering for hours what was going on =-}

So this patch instead adds a SmushListener interface, which is notified
at the time that _:foo is merged into _:bar, and makes Cursor
implement that interface and change the cursor position to _:bar
at that time.

darcs-hash:20050227155651-2ee97-7a7dc524c1dac79c1ab3620d27b21506d220d6dd