Change how spatial mainviews work: the lob they return is rendered full-screen.
Before, the lob returned by SpatialView.getMainviewLob() was supposed
to cover only part of the screen. Now, it will be rendered full-screen;
it can be an AlignLob wrapped around some other lob to render that
in the center of the screen.

The canvas view now makes the canvas full-screen,
which is an interesting effect and may help understanding
how the canvas is infinitely large...

darcs-hash:20050301225717-2ee97-4391da6038992210f29a5fdbea5a9571a69bb253