Full screen wasn't such a good idea, but *almost* full screen canvas looks quite good. :-)
The canvas view now has a 40-pixel margin around it; that makes it look
really nice.

Also added a lob that changes the depth fading, i.e. fog, to fade to the
canvas' background color instead of the scene's. This is nice because
most of the nodes overlap the canvas, but that's not the main reason;
the main reason is that we need to render the canvas bg behind everything,
but don't want to have it fade into the background -- argh. Need to find
a better method to do this... :-/

Currently it's really awful hackery...

darcs-hash:20050302002204-2ee97-2d28e9c738eccfdd9769769f71c02f8635808d69